proganonlaundryimages/requirements.txt
flask
ipython
matplotlib
opencv-python
pip-chill
tensorflow-gpu
